Changes innerText to innerHTML
Fixes an error of not parsing anything (replacing formulas for empty strings)
when the content of the formulas had not been properly parsed by the browser
(innerText requires parsing, innerHTML doesn't).

I did not investigate further to understand why .innerText did not work in some cases
(eg, when using formulas inside markdown files and then using bespoke-markdown to parse it).